requirements.txt
flask
flask-sqlalchemy
flask_script
pexpect
python3-nmap
flask-apscheduler